Moved plot_2d back into samples
